Refactor the code
Only support the following routes: /api/<id>, /api/spaces,
/api/create_user, /api/create_room, /api/check_in and
/api/check_out.

Don't give an error 500 when the db or fenix give an unexpected
response. In this case relay the message to the client.
Always give a 200 OK when processing a response even if the
response is not expected. Only give a 500 when a server function
crashes.