Add a new API endpoint called path
The path endpoint allows the user to browse the FenixEDU spaces API
with names in an hierarchical manner. For example
"api/path/alameda/pavilhao-central/0/0/GA2" will access the space
"alameda" looking for space "pavilhao-central". If the search is
successful it will then look for space "0", then space "0" and
finally "GA2".

Simplified the structs needed to serialize and deserialize the
resposnse needed. There is now only one struct "GenericSpace". The
parameter "capacity" is only filled when Fenix provides it,
otherwise it's empty.